requirements.txt
cffi==1.11.5
Django==1.11
django-bootstrap3==10.0.1
django-braces==1.13.0
misaka==2.1.0
Pillow==5.2.0
pycparser==2.18
pytz==2018.5